Move gateway::tests into a dedicated file
